requirements.txt
acres==0.2.0
aiofiles==24.1.0
annotated-types==0.7.0
azure-common==1.1.28
azure-core==1.32.0
azure-functions==1.21.3
azure-identity==1.20.0
azure-search-documents==11.5.2
azure-storage-blob==12.24.1
azure-storage-queue==12.22.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chardet==5.2.0
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
configobj==5.0.9
configparser==7.1.0
cryptography==43.0.3
distro==1.9.0
etelemetry==0.3.1
fastapi==0.115.8
filelock==3.16.1
frontend==0.0.3
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
iniconfig==2.0.0
isodate==0.6.1
itsdangerous==2.2.0
jiter==0.8.2
looseversion==1.3.0
lxml==5.3.1
msal==1.31.1
msal-extensions==1.2.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy==2.2.0
openai==1.57.2
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pluggy==1.5.0
portalocker==2.10.1
prov==2.0.1
puremagic==1.28
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydot==3.0.4
PyJWT==2.10.1
PyMuPDF==1.25.3
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
pytils==0.4.1
pytz==2025.1
pywin32; sys_platform == "win32"
pyxnat==1.6.3
rdflib==6.3.2
requests==2.32.3
scipy==1.15.2
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
tools==0.1.9
tqdm==4.67.1
traits==7.0.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
uvicorn==0.34.0
opencensus-ext-azure==1.1.14
